Move a chunk of logic into AvailableTimestamps
